+++ codex-rs/codex-agent/Cargo.toml
@@ -5,11 +5,25 @@ edition.workspace = true
 
 [dependencies]
 anyhow = { workspace = true }
-portable-pty = { workspace = true }
-tokio = { workspace = true, features = ["macros", "rt-multi-thread", "sync", "time"] }
+async-trait = { workspace = true }
 codex-protocol = { workspace = true }
+mcp-types = { workspace = true }
+base64 = { workspace = true }
+serde_json = { workspace = true }
+libc = { workspace = true }
+portable-pty = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
+sha1 = { workspace = true }
+shlex = { workspace = true }
+similar = { workspace = true }
+thiserror = { workspace = true }
+tokio = { workspace = true, features = ["macros", "process", "rt-multi-thread", "sync", "time"] }
+uuid = { workspace = true, features = ["serde", "v4"] }
+which = { workspace = true }
 wildmatch = { workspace = true }
+
+[dev-dependencies]
+core_test_support = { workspace = true }
 
 [lints]
 workspace = true
