+++ codex-rs/codex-agent/src/lib.rs
@@ -1,5 +1,19 @@
 pub mod config_types;
 pub mod exec_command;
+pub mod function_tool;
+pub mod notifications;
+pub mod services;
+pub mod shell;
+pub mod token_data;
 pub mod truncate;
+pub mod turn_diff_tracker;
+pub mod unified_exec;
 
 pub use config_types::*;
+pub use function_tool::*;
+pub use notifications::*;
+pub use services::*;
+pub use shell::*;
+pub use token_data::*;
+pub use turn_diff_tracker::*;
+pub use unified_exec::*;
