+++ codex-rs/codex-agent/src/unified_exec/errors.rs
