+++ codex-rs/core/src/codex.rs
@@ -39,14 +39,14 @@ use tracing::info;
 use tracing::trace;
 use tracing::warn;
 
+use crate::CredentialsProvider;
+use crate::McpInterface;
 use crate::ModelProviderInfo;
+use crate::Notifier;
+use crate::RolloutSink;
+use crate::SandboxManager;
 use crate::agent_config::AgentConfig;
-use crate::agent_services::CredentialsProvider;
 use crate::agent_services::DefaultSandboxManager;
-use crate::agent_services::McpInterface;
-use crate::agent_services::Notifier;
-use crate::agent_services::RolloutSink;
-use crate::agent_services::SandboxManager;
 use crate::apply_patch;
 use crate::apply_patch::ApplyPatchExec;
 use crate::apply_patch::InternalApplyPatchInvocation;
