+++ codex-rs/core/src/exec_command.rs
@@ -5,7 +5,6 @@ use crate::openai_tools::ResponsesApiTool;
 
 pub use codex_agent::exec_command::ExecCommandOutput;
 pub use codex_agent::exec_command::ExecCommandParams;
-pub use codex_agent::exec_command::ExecCommandSession;
 pub use codex_agent::exec_command::ExecSessionManager;
 pub use codex_agent::exec_command::WriteStdinParams;
