+++ codex-rs/core/src/lib.rs
@@ -95,12 +95,7 @@ pub use codex_protocol::protocol;
 pub use codex_protocol::config_types as protocol_config_types;
 
 pub use agent_config::AgentConfig;
-pub use agent_services::CredentialsProvider;
 pub use agent_services::DefaultSandboxManager;
-pub use agent_services::McpInterface;
-pub use agent_services::Notifier;
-pub use agent_services::RolloutSink;
-pub use agent_services::SandboxManager;
 pub use client::ModelClient;
 pub use client_common::Prompt;
 pub use client_common::REVIEW_PROMPT;
@@ -108,6 +103,12 @@ pub use client_common::ResponseEvent;
 pub use client_common::ResponseStream;
 pub use codex::compact::content_items_to_text;
 pub use codex::compact::is_session_prefix_message;
+pub use codex_agent::services::CredentialsProvider;
+pub use codex_agent::services::McpInterface;
+pub use codex_agent::services::Notifier;
+pub use codex_agent::services::ProviderAuth;
+pub use codex_agent::services::RolloutSink;
+pub use codex_agent::services::SandboxManager;
 pub use codex_protocol::models::ContentItem;
 pub use codex_protocol::models::LocalShellAction;
 pub use codex_protocol::models::LocalShellExecAction;
