+++ codex-rs/core/src/state/service.rs
@@ -1,7 +1,7 @@
-use crate::agent_services::McpInterface;
-use crate::agent_services::Notifier;
-use crate::agent_services::RolloutSink;
-use crate::agent_services::SandboxManager;
+use crate::McpInterface;
+use crate::Notifier;
+use crate::RolloutSink;
+use crate::SandboxManager;
 use std::sync::Arc;
 use tokio::sync::Mutex;
